requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.5.0
python-dotenv
prometheus-client
gunicorn
structlog